Initial set of functions
Working on abstractions for polynomials, splines, NURBS, polygons,  2D Quadrics (circle and ellipse). Each of these is written as a MATLAB "classdef" (which is the closest feature that resembles a C struct in MATLAB). 
Some functions are standalone and will be merged into the corresponding classdef definitions as this project matures.

MATLAB is being used for its handy plotting utilities (why not Python? Well, my university offers free student licenses for MATLAB which I will use for the time being). Need to learn OpenGL or Vulkan to replicate the plotting functionality. With that being said, the functions are being coded in a very "primal" way. The use of high-level functions is by design limited. If a high-level function is present that is because I have not yet figured out how to replicate its functionality myself.